Moved LoginRoute and associated files to respective directory
