+++ src/ClientRoute/LoginRoute/LoginRoute.jsx
@@ -1,13 +1,12 @@
 import React, { useEffect } from "react";
 import { useRef } from "react";
 import { Navigate, Outlet } from "react-router-dom";
-
-import { useNotification } from "../Notification";
+import { useNotification } from "../../Notification";
 
 export default function LoginRoute(props) {
   const { isConnected } = props;
   const initalConnection = useRef();
-  const { successAlert, failedAlert } = useNotification();
+  const { successAlert } = useNotification();
 
   useEffect(() => {
     if (isConnected) {
